+++ ai_config/ai_constants.py
@@ -40,8 +40,5 @@ MMR_SEARCH_K = 5
 MMR_FETCH_K = 8
 MMR_LAMBDA_MULT = 0.8
 
-# URL we send the payload to for translation
-URL_TRANSLATE_EN_TO_MS = "https://api.mesolitica.com/translation"
-
 # Language Default
 LANGUAGE_DEFAULT = "en"
