+++ api/chatbot.py
@@ -1,9 +1,7 @@
-import gc
 import json
 import logging
 import os
 import time
-import tracemalloc
 
 from typing import List
 
@@ -39,13 +37,7 @@ from ai_config.ai_prompts import (
     RAG_PROMPT_TEMPLATE,
 )
 
-from api.views.brain_file_reader import (
-    get_document_by_id,
-    get_document_by_question_text,
-    update_answer_detailed,
-    insert_new_document,
-    get_next_id_from_json
-)
+
 
 from api.brain_retriever import (
   MultiRetriever,
@@ -58,20 +50,6 @@ from .ai_services import (
   GradeConfidenceLevel,
  )
 
-def monitor_memory():
-    """Start memory monitoring and return initial snapshot"""
-    tracemalloc.start()
-    snapshot1 = tracemalloc.take_snapshot()
-    return snapshot1
-
-def compare_memory(snapshot1):
-    """Compare memory usage against initial snapshot"""
-    snapshot2 = tracemalloc.take_snapshot()
-    top_stats = snapshot2.compare_to(snapshot1, "lineno")
-    print("[ Top 10 memory differences ]")
-    for stat in top_stats[:10]:
-        print(stat)
-    gc.collect()  # Force garbage collection after comparison
 
 # create our gloabl variables
 logger = logging.getLogger(__name__)
@@ -371,310 +349,3 @@ def handle_mongodb_operation(operation):
         return None
 
 
-
-def extrair_knowledge_items(conversation):
-    """
-    Analisa a conversa, extrai os pontos de resolução e cria um resumo conciso desses pontos.
-    Retorna um item de conhecimento multilíngue com um resumo desses trechos.
-    """
-    # Concatene todas as mensagens da conversa (ou selecione apenas as que podem conter resolução)
-    full_text = "\n".join(
-        msg.get("content", "") for msg in conversation.get("messages", []) if msg.get("content")
-    )
-    
-    extraction_prompt = (
-        "Analyze the conversation below and extract the parts where the user's problem was solved "
-        "or the issue was resolved. Provide a concise summary in English, capturing the resolutions or "
-        "key decisions made. Do not include any extraneous details. Provide the full resolution details first, "
-        "followed by a concise, reduced summary. "
-        "If the conversation does not contain a resolution regarding a real problem, please state in English that "
-        "there is no relevant resolution in this conversation.\n\nConversation:\n" + full_text
-    )   
-    
-    api_key = os.getenv("OPENAI_API_KEY")
-    if not api_key:
-        raise Exception("OPENAI_API_KEY is missing")
-    
-    client = OpenAI(api_key=api_key)
-    try:
-        response = client.chat.completions.create(
-            model=OPENAI_MODEL,
-            messages=[
-                {"role": "system", "content": "You are a helpful assistant that extracts and summarizes resolution points from a conversation."},
-                {"role": "user", "content": extraction_prompt},
-            ],
-            temperature=0.2,
-            max_tokens=130,
-            timeout=OPENAI_TIMEOUT,
-        )
-    except Exception as e:
-        logger.error(f"Error during resolution extraction and summary generation: {str(e)}")
-        return []
-
-    extracted_text = response.choices[0].message.content.strip()
-    
-    if not extracted_text or extracted_text.lower() == "none":
-        return []
-
-    if "Summary:" in extracted_text:
-        full_resolution, reduced_summary = extracted_text.split("Summary:", 1)
-        reduced_summary = reduced_summary.strip()
-    else:
-        full_resolution = extracted_text
-        reduced_summary = full_resolution
-   
-    full_resolution = full_resolution.replace("Resolution:", "").strip()
-    
-    candidate_item = {
-        "id": conversation["session_id"] + "_resolucao",
-        "question": {
-            "text": "What are the resolution points of the conversation?", 
-            "variations": [],
-            "intent": "extracted_resolution_points",
-            "languages": {
-                "en": "Resolution points of the conversation",
-            }
-        },
-        "answer": {
-           # "short": {
-            #    "en": reduced_summary,
-            # },
-            "detailed": {
-                "en": full_resolution,
-            },
-            "conditions": ["Derived from conversation resolution"]
-        },
-        "metadata": {
-            "category": ["extracted"],
-            "subCategory": "conversation_resolution",
-            "confidence": 0.7,
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "followUpQuestions": {}
-        }
-    }
-    
-    try:
-        docs_retrieve = retriever.get_relevant_documents(extracted_text)
-        logger.debug(f"Retrieved {len(docs_retrieve)} documents.")
-        
-        docs_to_use = []
-        reasoning = []
-        seen_ids = set()
-        
-        doc_ids = []
-        for doc in docs_retrieve:
-            doc_id = doc.metadata.get('id', 'no_id')
-            relevance_score = retrieval_grader.invoke({
-                "prompt": extracted_text, 
-                "document": doc.page_content
-            })
-            if relevance_score.confidence_score >= 0.7:
-                doc_id = doc.metadata.get('id', 'no_id')
-                if doc_id == "no_id":
-                    question_part = doc.page_content.split("Answer:")[0]
-                    question_only = question_part.replace("Question:", "").strip()
-                    doc_id=get_document_by_question_text(question_only)
-                    
-                if doc_id not in seen_ids:
-                    doc_ids.append(doc_id)
-                    seen_ids.add(doc_id)
-                    docs_to_use.append(doc)
-                    reasoning.append({
-                        "id": doc_id,
-                        "document": doc.page_content,
-                        "score": relevance_score.confidence_score,
-                        "rationale": f"Document (ID: {doc_id}) is relevant with score {relevance_score.confidence_score}."
-                    })
-
-        context = format_docs(docs_to_use)
-
-        response = rag_chain.invoke({
-            "context": context,
-            "prompt": extracted_text,
-            "reasoning": reasoning
-        })
-
-        logger.debug("Response generated using the following documents:")
-        responseRaw = []
-        for r in reasoning:
-            logger.debug(f"Rationale: {r['rationale']}\nDocument Content: {r['document'][:200]}...")
-
-        valid_doc_ids = [doc_id for doc_id in doc_ids if doc_id != "no_id"]
-
-        if valid_doc_ids:
-            for ID in valid_doc_ids:
-                doc = get_document_by_id(ID) 
-            if doc:
-                responseRaw.append(json.dumps(doc, indent=4, ensure_ascii=False))
-            
-        candidate_item["answer"]["rag_response"] = response
-        candidate_item["answer"]["reasoning"] = reasoning
-        candidate_item["answer"]["raw"] = (responseRaw)
-
-        return [candidate_item]
-
-    except Exception as ve:
-        logger.error(f"Error during vector store retrieval: {str(ve)}")
-        return []
-    
-    
-def search_by_id(store: Chroma, custom_id: str):
-    results = store.get(
-        where={"id": custom_id},
-        include=["documents", "metadatas"]
-    )
-    return results
-
-def update_document_by_custom_id(custom_id: str, answer_en: str, answer_ms: str, answer_cn: str):
-    try:
-        search_results = store.get(
-            where={"id": custom_id},
-            include=["metadatas","documents"]
-        )
-
-        if not search_results['ids']:
-            print(f"Document ID '{custom_id}' not found.")
-            return
-
-        document_id = search_results['ids'][0]
-        doc = get_document_by_id(document_id) 
-        
-        if doc:
-            update_answer_detailed(doc, answer_en, answer_ms, answer_cn)
-
-        existing_metadata = search_results['metadatas'][0]
-        document = search_results['documents'][0]
-        question_text = document.split("\n")[0].replace("Question: ", "").strip()
-
-        new_document = CustomDocument(
-            id=custom_id,
-            page_content=(
-                f"Question: {question_text}\n"
-                f"Answer: {answer_en}"
-            ),
-            metadata={
-                "id": doc.get("id", ""),
-                "category": ",".join(existing_metadata.get("category", [])),
-                "subCategory": existing_metadata.get("subCategory", ""),
-                "difficulty": existing_metadata.get("difficulty", 0),
-                "confidence": existing_metadata.get("confidence", 0.0),
-                "intent": doc["question"].get("intent", ""),
-                "variations": ", ".join(doc["question"].get("variations", [])),
-                "conditions": ", ".join(doc["answer"].get("conditions", [])),
-            },
-        )
-
-        store.add(documents=[new_document])
-        return (f"ID '{custom_id}' updated ")
-    except Exception as e:
-        print(f"Error to update document: {str(e)}")
-
-def reload_vector_store(collection_name="RAG", persist_directory="./chroma_db", embedding_model=None):
-    return Chroma(
-        collection_name=collection_name,
-        persist_directory=persist_directory,
-        embedding_function=embedding_model
-    )
-         
-def insert_document(question_text, answer_en: str, answer_ms: str, answer_cn: str):
-    try:
-        
-        new_id = get_next_id_from_json()
-        if not new_id:
-            return "Error obtaining the next ID."
-    
-        new_document = {
-            "id": new_id,
-            "question": {
-            "text": question_text,
-            "variations": [],
-            "intent": "general_inquiry",
-            "languages": {
-                "en": "",
-                "ms": "",
-                "cn": ""
-            }
-        },
-        "answer": {
-            "detailed": {
-                "en": answer_en,
-                "ms": answer_ms,
-                "cn": answer_cn
-            },
-            "conditions": []
-        },
-        "metadata": {
-            "category": ["general"],
-            "subCategory": "general_inquiry",
-            "difficulty": 0,
-            "confidence": 0.5,
-            "dateCreated": "",
-            "lastUpdated": "",
-            "version": "1.0",
-            "source": "",
-            "status": "active"
-        },
-        "context": {
-            "relatedTopics": [],
-            "prerequisites": [],
-            "followUpQuestions": {
-                "en": [],
-                "ms": [],
-                "cn": []
-            }
-        },
-        "usage": {
-            "searchFrequency": 0,
-            "successRate": 0,
-            "lastQueried": None
-        },
-        "review_status": []
-        }
-
-        insert_new_document(new_document)
-
-        new_document = CustomDocument(
-            id=new_id,
-            page_content=(
-                f"Question: {question_text}\n"
-                f"Answer: {answer_en}"
-            ),
-            metadata={
-                "id": new_id,
-                "category": ",".join(["general"]),
-                "subCategory": "general_inquiry",
-                "difficulty":  0,
-                "confidence": 0.5,
-                "intent": "general_inquiry",
-                "variations": ",".join(["general"]),
-                "conditions": ",".join(["general"])
-            },
-        )
-
-        store.add_documents(documents=[new_document])
-        store.persist()
-        return (f"ID '{new_id}' updated ")
-    
-
-        #store.add(documents=[new_document])
-        #store.add_documents(documents=[new_document])
-       # 
-
-        #collection_name="RAG",
-        #embedding=embedding_model,
-        #persist_directory="./chroma_db",
-        
-        #store = Chroma(
-         #   collection_name=collection_name,
-          #  persist_directory=persist_directory,
-           # embedding_function=embedding
-        #)
-
-        #return (f"ID '{new_id}' updated ")
-    except Exception as e:
-        print(f"Error to update document: {str(e)}")
-
-
