+++ data/database_part_1.json
@@ -9330,7 +9330,7 @@
   {
     "id": "0177",
     "question": {
-      "text": "Boss, I want to find Lucky Spin",
+      "text": "Boss, I want to play Lucky Spin",
       "variations": [
         "Bos, saya nak cari Lucky Spin",
         "Boss, I want to find Lucky Spin",
